// frontend/src/app/core/api/model/publicMediaAssetResponse.ts
/**
 * Le Tacos Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PublicMediaAssetResponseDto { 
    id?: string;
    url?: string;
    alt?: string;
    primary?: boolean;
    displayOrder?: number;
}