// frontend/src/app/core/api/model/publicMenuCategoryGroupResponse.ts
/**
 * Le Tacos Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PublicCategoryResponseDto } from './publicCategoryResponse';
import { PublicMenuItemResponseDto } from './publicMenuItemResponse';


export interface PublicMenuCategoryGroupResponseDto { 
    category?: PublicCategoryResponseDto;
    items?: Array<PublicMenuItemResponseDto>;
}